@@ -15,16 +15,17 @@
   "license": "ISC",
   "devDependencies": {
     "@types/escape-html": "^1.0.4",
-    "@types/node": "^20.14.0",
-    "autoprefixer": "^10.4.19",
+    "@types/node": "^20.17.6",
+    "@vue/tsconfig": "^0.5.1",
+    "autoprefixer": "^10.4.20",
     "escape-html": "^1.0.3",
     "markdown-it-task-lists": "^2.1.1",
-    "pagefind": "^1.1.0",
-    "postcss": "^8.4.38",
-    "postcss-preset-env": "^9.5.14",
-    "tailwindcss": "^3.4.3",
-    "vitepress": "^1.2.2",
-    "vitepress-plugin-pagefind": "^0.4.0",
-    "vue": "^3.4.27"
+    "pagefind": "^1.2.0",
+    "postcss": "^8.4.49",
+    "postcss-preset-env": "^9.6.0",
+    "tailwindcss": "^3.4.15",
+    "vitepress": "^1.5.0",
+    "vitepress-plugin-pagefind": "^0.4.11",
+    "vue": "^3.5.13"
   }
 }
